Move MatrixChat to react-sdk. Move all login stuff to react SDK.
Removed Modulator stuff.
