+++ src/VectorConferenceHandler.js
